movement to beat
