+++ Assets/Scripts/BeatManager.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: d7eec1d059f983b4b8a5d9aa6223cfd0
