+++ Assets/Scripts/PlayerController.cs
@@ -1,28 +1,5 @@
 using UnityEngine;
 using System.Collections.Generic;
-using UnityEngine.Events;
-
-[System.Serializable]
-public class Intervals
-{
-     [SerializeField] private float steps;
-     [SerializeField] private UnityEvent trigger;
-     private int lastInterval;
-
-     public float GetIntervalLength(float bpm)
-     {
-          return 60f / (bpm * steps);
-     }
-
-     public void CheckForNewInterval (float interval)
-     {
-          if (Mathf.FloorToInt(interval) != lastInterval)
-          {
-               lastInterval = Mathf.FloorToInt(interval);
-               trigger.Invoke();
-          }
-     }
-}
 
 public class PlayerController2D : MonoBehaviour
 {
@@ -43,24 +20,13 @@ public class PlayerController2D : MonoBehaviour
 
      Vector2 direction; // Stores the direction of player movement
      
-     
-
-     // Private variables 
-     private bool isMovingHorizontally = true; // Flag to track if the player is moving horizontally
-
-     [SerializeField] private float bpm;
-     [SerializeField] private AudioSource music; // Handles synchronization of the music
-     [SerializeField] private Intervals[] intervals;
-
      void Start()
      {
          
      }
 
-     void Update()
+     public void Update()
      {
-          Synchronize(music);
-
           // get direction of input
           direction = new Vector2(Input.GetAxis("Horizontal"), Input.GetAxis("Vertical")).normalized;
 
@@ -69,6 +35,7 @@ public class PlayerController2D : MonoBehaviour
 
           // get sprite that faces same direction as input
           List<Sprite> directionSprites =  GetSpriteDirection();
+
           if (directionSprites != null )
           {
                spriteRenderer.sprite = directionSprites[0];
@@ -100,15 +67,6 @@ public class PlayerController2D : MonoBehaviour
           return selectedSprites;
      }
      
-     void Synchronize(AudioSource music)
-     {
-          foreach (Intervals interval in intervals)
-          {
-               float sampledTime = (music.timeSamples / music.clip.frequency * interval.GetIntervalLength(bpm));
-               interval.CheckForNewInterval(sampledTime);
-
-          }
-
-     }
+     
 
 }
