+++ Packages/manifest.json
@@ -5,7 +5,7 @@
     "com.unity.feature.2d": "2.0.1",
     "com.unity.ide.rider": "3.0.31",
     "com.unity.ide.visualstudio": "2.0.22",
-    "com.unity.ide.vscode": "1.2.3",
+    "com.unity.ide.vscode": "1.2.5",
     "com.unity.inputsystem": "1.13.0",
     "com.unity.multiplayer.center": "1.0.0",
     "com.unity.render-pipelines.universal": "17.0.3",
